refactor v1 stuff
