+++ internal/handlers/v1handlers/deploys_handlers.go
@@ -2,13 +2,12 @@ package v1handlers
 
 import (
 	"errors"
-	"github.com/broadinstitute/sherlock/internal/config"
 	"github.com/broadinstitute/sherlock/internal/controllers/v1controllers"
 	"github.com/broadinstitute/sherlock/internal/models/v1models"
 	"github.com/broadinstitute/sherlock/internal/serializers/v1serializers"
 	"net/http"
 
-	"github.com/broadinstitute/sherlock/internal/metrics"
+	"github.com/broadinstitute/sherlock/internal/metrics/v1metrics"
 	"github.com/gin-gonic/gin"
 )
 
@@ -92,15 +91,13 @@ func createDeploy(dc *v1controllers.DeployController) func(c *gin.Context) {
 			return
 		}
 
-		if config.Config.String("metrics.accelerate.fromAPI") != "v2" {
-			metrics.RecordDeployFrequency(c, environment, service)
-			// calculate lead time
-			if shouldUpdateLeadTime {
-				leadTime := deploy.CalculateLeadTimeHours()
-				metrics.RecordLeadTime(c, leadTime, environment, service)
-			}
+		v1metrics.RecordDeployFrequency(c, environment, service)
+		// calculate lead time
+		if shouldUpdateLeadTime {
+			leadTime := deploy.CalculateLeadTimeHours()
+			v1metrics.RecordLeadTime(c, leadTime, environment, service)
 		}
-
+		
 		c.JSON(http.StatusCreated, v1serializers.DeploysResponse{Deploys: dc.Serialize(deploy)})
 	}
 }
